Added names for Delphi releases before XE.
Although Delphinus only supports XE and newer, it is perfectly valid
scenario that a package could support older Delphis. (For example,
OmniThreadLibrary supports Delphi 2007 and newer.) It is therefor
desired that correct product name is displayed in the Package Info
dialog in that case (and not 'Delphi Compiler 19' or similar).